+++ src/Controller/LogController.php
@@ -8,8 +8,8 @@ use AnzuSystems\CommonBundle\ApiFilter\ApiParams;
 use AnzuSystems\CommonBundle\Document\Log;
 use AnzuSystems\CommonBundle\Log\LogFacade;
 use AnzuSystems\CommonBundle\Log\Model\LogDto;
-use AnzuSystems\CommonBundle\Log\Repository\AppLogRepository;
 use AnzuSystems\CommonBundle\Log\Repository\AuditLogRepository;
+use AnzuSystems\CommonBundle\Log\Repository\JournalLogRepository;
 use AnzuSystems\CommonBundle\Model\OpenApi\Parameter\OAParameterPath;
 use AnzuSystems\CommonBundle\Model\OpenApi\Request\OARequest;
 use AnzuSystems\CommonBundle\Model\OpenApi\Response\OAResponse;
@@ -26,8 +26,8 @@ use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 final class LogController extends AbstractAnzuApiController
 {
     public function __construct(
-        private readonly ?AuditLogRepository $auditLogRepo,
-        private readonly ?AppLogRepository $appLogRepo,
+        private readonly AuditLogRepository $auditLogRepo,
+        private readonly JournalLogRepository $journalLogRepo,
         private readonly LogFacade $logFacade,
     ) {
     }
@@ -40,29 +40,21 @@ final class LogController extends AbstractAnzuApiController
     #[OAResponseList(Log::class)]
     public function getAuditLogs(ApiParams $apiParams): JsonResponse
     {
-        if (null === $this->auditLogRepo) {
-            throw new NotFoundHttpException('Not configured to serve audit logs.');
-        }
-
         return $this->okResponse(
             $this->auditLogRepo->findByApiParams($apiParams)
         );
     }
 
     /**
-     * Get list of app logs.
+     * Get list of journal logs.
      *
      * @throws SerializerException
      */
     #[OAResponseList(Log::class)]
-    public function getAppLogs(ApiParams $apiParams): JsonResponse
+    public function getJournalLogs(ApiParams $apiParams): JsonResponse
     {
-        if (null === $this->appLogRepo) {
-            throw new NotFoundHttpException('Not configured to serve app logs.');
-        }
-
         return $this->okResponse(
-            $this->appLogRepo->findByApiParams($apiParams)
+            $this->journalLogRepo->findByApiParams($apiParams)
         );
     }
 
@@ -70,18 +62,14 @@ final class LogController extends AbstractAnzuApiController
      * Get one app log.
      */
     #[OAParameterPath('id'), OAResponse(Log::class)]
-    public function getOneAppLog(string $id): JsonResponse
+    public function getOneJournalLog(string $id): JsonResponse
     {
-        if (null === $this->appLogRepo) {
-            throw new NotFoundHttpException('Not configured to serve app logs.');
-        }
-
-        $log = $this->appLogRepo->find($id);
+        $log = $this->journalLogRepo->find($id);
         if ($log instanceof Log) {
             return $this->okResponse($log);
         }
 
-        throw new NotFoundHttpException(sprintf('App log with id "%s" was not found.', $id));
+        throw new NotFoundHttpException(sprintf('Journal log with id "%s" was not found.', $id));
     }
 
     /**
@@ -92,10 +80,6 @@ final class LogController extends AbstractAnzuApiController
     #[OAParameterPath('id'), OAResponse(Log::class)]
     public function getOneAuditLog(string $id): JsonResponse
     {
-        if (null === $this->auditLogRepo) {
-            throw new NotFoundHttpException('Not configured to serve audit logs.');
-        }
-
         $log = $this->auditLogRepo->find($id);
         if ($log instanceof Log) {
             return $this->okResponse($log);
@@ -105,7 +89,7 @@ final class LogController extends AbstractAnzuApiController
     }
 
     /**
-     * Create custom app log (i.e. Admin FE error log).
+     * Create custom journal log (i.e. Admin FE error log).
      *
      * @throws SerializerException
      */