+++ CHANGELOG.md
@@ -1,3 +1,57 @@
+## [10.0.0](https://github.com/anzusystems/common-bundle/compare/7.0.0...8.0.0) (2024-10-30)
+
+### Features
+ * App logger is the default logger expected to be logged to Sentry or Syslog and log errors, not mongodb. To keep logging some other data in Mongodb, new journal logger was created. 
+
+### Changes
+* `anzu_mongo_app_log_collection` was renamed to `anzu_mongo_journal_log_collection`
+* Routing changes - `app` was renamed to `journal`
+* BC change: configuration change
+
+Before:
+```php
+    $logsConfig
+        ->app()
+            ->ignoredExceptions([
+                NotFoundHttpException::class,
+                AccessDeniedException::class,
+                UnauthorizedHttpException::class,
+                ValidationException::class,
+            ])
+            ->mongo()
+                ->uri(env('ANZU_MONGODB_APP_LOG_URI'))
+                ->username(env('ANZU_MONGODB_APP_LOG_USERNAME'))
+                ->password(env('ANZU_MONGODB_APP_LOG_PASSWORD'))
+                ->database(env('ANZU_MONGODB_APP_LOG_DB'))
+                ->ssl(env('ANZU_MONGODB_APP_LOG_SSL')->bool())
+                ->collection('appLogs')
+    ;
+```
+
+Now:
+```php
+    $logsConfig
+        ->app()
+            ->ignoredExceptions([
+                NotFoundHttpException::class,
+                AccessDeniedException::class,
+                UnauthorizedHttpException::class,
+                ValidationException::class,
+            ])
+    ;
+
+    $logsConfig
+        ->journal()
+            ->mongo()
+                ->uri(env('ANZU_MONGODB_APP_LOG_URI'))
+                ->username(env('ANZU_MONGODB_APP_LOG_USERNAME'))
+                ->password(env('ANZU_MONGODB_APP_LOG_PASSWORD'))
+                ->database(env('ANZU_MONGODB_APP_LOG_DB'))
+                ->ssl(env('ANZU_MONGODB_APP_LOG_SSL')->bool())
+                ->collection('appLogs')
+    ;
+```
+
 ## [8.0.0](https://github.com/anzusystems/common-bundle/compare/7.0.0...8.0.0) (2024-05-29)
 ### Features
 * Added command `anzusystems:user:sync-base` for loading basic user set (depends on `user_sync_data` configuration)